pyproject.toml
[project]
name = "ai-front-desk"
version = "0.1.0"
description = "AI Front Desk — intelligent chat assistant for daycare centers"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
    "litellm>=1.83.0",
    "langgraph>=1.1.0",
    "langchain-core>=1.2.0",
    "langsmith>=0.7.0",
    "aiosqlite>=0.22.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=3.3.0",
    "python-frontmatter>=1.1.0",
    "scikit-learn>=1.8.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.15.0",
]
documents = [
    "docling>=2.18.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
include = ["backend*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
